Bug 24117.        Initial stab at refactoring to support multi-site detection.        Pull out mixture estimation methods into a common class;  previously there were copies of methods in KineticWorker and ModificationDecode.
[git-p4: depot-paths = "//depot/software/bioinformatics/tools/kineticsTools/": change = 128739]
