+++ src/pbtools/kineticsTools/ModificationDecode.py
@@ -38,97 +38,28 @@ from scipy.special import gammaln as gamln
 from numpy import log, pi, log10, e, log1p, exp
 import numpy as np
 
-
-log10e = log10(e)
-
-canonicalBaseMap = { 'A': 'A', 'C':'C', 'G':'G', 'T':'T', 'H':'A', 'I':'C', 'J':'C', 'K':'C' }
-modNames = { 'H':'m6A', 'I':'m5C', 'J':'m4C', 'K':'m5C' }
-
-ModificationPeakMask = { 'm6A' : [0, -5], 'm4C': [0, -5], 'm5C': [2, 0, -1, -2, -4, -5, -6]  }
-
-# Labels for modified fraction:
-
-FRAC = 'frac'
-FRAClow = 'fracLow'
-FRACup = 'fracUp'
-
-# Try computing these only once
-
-k1 = s.norm.ppf(0.025)
-k2 = s.norm.ppf(0.975)
+from MultiSiteCommon import MultiSiteCommon, canonicalBaseMap, modNames, ModificationPeakMask, FRAC, FRAClow, FRACup, log10e
+from MixtureEstimationMethods import MixtureEstimationMethods
 
 
-class ModificationDecode(object):
 
-    def __init__(self, gbmModel, sequence, rawKinetics, callBounds, methylMinCov, modsToCall = ['H', 'J', 'K'], methylFractionFlag = False, useLDAFlag = False):
-        """
-        All indexes are 0-based into the the sequence.
+class ModificationDecode( MultiSiteCommon ):
 
-        find a set of sites that _might_ have a modification - each modification type will include a list of
-        'neighbor peaks' that can add the current site to the 'options' list.
-        6mA and 4mC will use only the on-target peak
-        5caC will use on target, -2 and -6.
+    def __init__(self, gbmModel, sequence, rawKinetics, callBounds, methylMinCov, modsToCall = ['H', 'J', 'K'], methylFractionFlag = False, useLDAFlag = False ):
 
-        Only hits that make this list will be tested in the mod identification process
-
-        Use the viterbi algorithm to find the optimal modifications to include, by measuring the per-site likelihood
-        of the observed IPD, given the underlying sequence and methylation states.
-        """
-
-        self.methylMinCov = methylMinCov
-
-        # Temporary:
-        self.useLDA = useLDAFlag
-
-        self.modsToCall = modsToCall
-
-        self.methylFractionFlag = methylFractionFlag	
-   
-        log1p = math.log(0.05)
-        self.modPriors = { 'H': log1p, 'I': log1p, 'J': log1p, 'K': log1p }
-
-        self.gbmModel = gbmModel
-        self.sequence = sequence
-
-        self.callStart = callBounds[0]
-        self.callEnd = callBounds[1]
+        MultiSiteCommon.__init__(self, gbmModel, sequence, rawKinetics)
 
         # Extents that we will attemp to call a modification
+        self.callStart = callBounds[0]
+        self.callEnd = callBounds[1]
         self.callRange = xrange(self.callStart, self.callEnd)
 
-        # These switch because we changing viewpoints
-        self.pre = gbmModel.post
-        self.post = gbmModel.pre
-
-        self.lStart = self.pre
-        self.lEnd = len(self.sequence) - self.post
-
-        # Extents that we will use for likelihoods
-        self.likelihoodRange = xrange(self.lStart, self.lEnd)
-
-        self.alternateBases = dict((x, set(sequence[x])) for x in xrange(len(sequence)))
+        self.methylMinCov = methylMinCov
+        self.modsToCall = modsToCall
+        self.methylFractionFlag = methylFractionFlag
+        self.useLDA = useLDAFlag        
 
 
-        self.rawKinetics = rawKinetics
-
-
-    def getConfigs(self, centerIdx):
-        start = centerIdx - self.pre
-        end = centerIdx + self.post
-        return self._possibleConfigs(start, end)
-
-    def _possibleConfigs(self, start, end):
-
-        if start == end:
-            return self.alternateBases[start]
-        else:
-            r = []
-            currentChars = self.alternateBases[start]
-            for suffix in self._possibleConfigs(start+1, end):
-                for c in currentChars:
-                    r.append(c + suffix)
-
-            return r
 
     def decode(self):
         """Use this method to do the full modification finding protocol"""
@@ -160,11 +91,11 @@ class ModificationDecode(object):
         for (pos, peak) in self.rawKinetics.items():
             score = peak['score']
 
-            if self.useLDA:
-                # Try using LDA model to identify putative Ca5C, regardless of scores
-                if peak.has_key('Ca5C'):
-                    if peak['Ca5C'] < 0:
-                        self.alternateBases[pos].add('K')
+            # if self.useLDA:
+            #     # Try using LDA model to identify putative Ca5C, regardless of scores
+            #     if peak.has_key('Ca5C'):
+            #         if peak['Ca5C'] < 0:
+            #             self.alternateBases[pos].add('K')
 
             # Exclude points with low score
             if score < scoreThresholdLow:
@@ -205,75 +136,6 @@ class ModificationDecode(object):
                     if seq[pos + 6] == 'C' and pos + 6 < self.callEnd:
                         self.alternateBases[pos + 6].add('K')
 
-
-
-    def computeContextMeans(self):
-        """Generate a hash of the mean ipd for all candidate contexts"""
-        allContexts = list(set([ cfg for pos in self.likelihoodRange for cfg in self.getConfigs(pos) ]))
-        predictions = self.gbmModel.getPredictions(allContexts)
-        self.contextMeanTable = dict(zip(allContexts, predictions))
-
-
-    # Log-t pdf - copied from scipy distributions.py line 3836
-
-    def _logpdf(self, x, df):
-        r = df*1.0
-        lPx = gamln((r+1)/2)-gamln(r/2)
-        lPx -= 0.5*log(r*pi) + (r+1)/2*log(1+(x**2)/r)
-        return lPx
-
-
-    def singleScore(self, position, context):
-        if self.rawKinetics.has_key(position):
-            siteObs = self.rawKinetics[ position ]
-
-            # mu of model, error in model
-            um = self.contextMeanTable[context]
-
-            # FIXME -- unify this with the error model used in KineticWorker.py
-            # em = 0.06 * um + 0.12 * um**2.0
-            em = 0.01 + 0.03*um + 0.06*um**(1.7)
-
-            uo = siteObs['tMean']
-            eo = siteObs['tErr']
-
-            t = -(uo - um) / sqrt(em**2 + eo**2)
-            df = max(1, siteObs['coverage'] - 1)
-
-            logLikelihood = self._logpdf(t, df).item()
-            # logLikelihood = s.t.logpdf(t, df).item()
-        else:
-            logLikelihood = 0
-        
-        return logLikelihood
-
-
-    def scorePosition(self, position, context):
-        """ Compute the likelihood of the observed IPDs at position, given the context"""
-
-        # Handle the prior for a modification at the current base here
-        # unmodified bases get a prior of 0, modified bases get a prior less than 0.
-        prior = 0.0
-        if self.modPriors.has_key(context[self.pre]):
-            prior = self.modPriors[context[self.pre]]
-
-        # Handle positions where we don't have enough coverage
-        if not self.rawKinetics.has_key( position ):
-            return prior
-
-        # FIXME:  temporary change to try LDA for Ca5C detection?       
-        # if self.useLDA:
-        #     # sum scores over a window around current position:
-        #     ll = 0
-        #     for offset in xrange(-3, 3):
-        #         ll += self.singleScore(position+offset, context) 
-        # else:
-        #     ll = self.singleScore(position, context)
-        # Doesn't seem to work?
-
-        ll = self.singleScore( position, context )
-        # return logLikelihood.item() + prior
-        return ll + prior
 
 
     def fwdRecursion(self):
@@ -356,53 +218,53 @@ class ModificationDecode(object):
                 break
 
 
-        if self.useLDA:
-            # allow LDA-predicted sites through to GFF file
-            for pos in range(start, end):
-                if self.rawKinetics.has_key(pos):
-                    if self.rawKinetics[pos].has_key('Ca5C'):
-                        if 'K' in self.modsToCall:
-                            # cutoff = min( 0, self.rawKinetics[pos]['coverage']/20.0 - 3.0 )
-                            cutoff = 0
-                            echoSites = [modCalls[pos + i] for i in [-6, -5, -2, -1, 2] if modCalls.has_key(pos + i)]
-                        else:
-                            cutoff = -2.25
-                            echoSites = [modCalls[pos + i] for i in range(-10,11) if modCalls.has_key(pos + i)]
-                        if self.rawKinetics[pos]['Ca5C'] < cutoff:
-                            # so long as those sites are not in the vicinity of a m6A/m4C call
-                            if 'H' not in echoSites and 'J' not in echoSites:
-                                modCalls[pos] = 'K'
-                        # else:
-                        #     # remove any non-LDA-predicted sites from modCalls dictionary?
-                        #     if modCalls.has_key(pos):
-                        #         del modCalls[pos]
+        # if self.useLDA:
+        #     # allow LDA-predicted sites through to GFF file
+        #     for pos in range(start, end):
+        #         if self.rawKinetics.has_key(pos):
+        #             if self.rawKinetics[pos].has_key('Ca5C'):
+        #                 if 'K' in self.modsToCall:
+        #                     # cutoff = min( 0, self.rawKinetics[pos]['coverage']/20.0 - 3.0 )
+        #                     cutoff = 0
+        #                     echoSites = [modCalls[pos + i] for i in [-6, -5, -2, -1, 2] if modCalls.has_key(pos + i)]
+        #                 else:
+        #                     cutoff = -2.25
+        #                     echoSites = [modCalls[pos + i] for i in range(-10,11) if modCalls.has_key(pos + i)]
+        #                 if self.rawKinetics[pos]['Ca5C'] < cutoff:
+        #                     # so long as those sites are not in the vicinity of a m6A/m4C call
+        #                     if 'H' not in echoSites and 'J' not in echoSites:
+        #                         modCalls[pos] = 'K'
+        #                 # else:
+        #                 #     # remove any non-LDA-predicted sites from modCalls dictionary?
+        #                 #     if modCalls.has_key(pos):
+        #                 #         del modCalls[pos]
 
 
           
         # correct adjacent calls:
-        if self.useLDA:
-            for pos in range(start + 2, end - 2, 2 ):
-                x = [pos + i for i in range(-2,3) if modCalls.has_key( pos + i) and self.rawKinetics.has_key( pos + i) ]
-                y = [modCalls[j] for j in x]
-                if y.count('K') > 1:
-                    tmp = [self.rawKinetics[j]['Ca5C'] for j in x if self.rawKinetics[j].has_key('Ca5C') ]
-                    if len(tmp) > 0:
-                        lowest = min(tmp)
-                        for j in x:
-                            if self.rawKinetics[j].has_key('Ca5C'):
-                                if self.rawKinetics[j]['Ca5C'] > lowest:
-                                    del modCalls[j]
-                    
+        # if self.useLDA:
+        #     for pos in range(start + 2, end - 2, 2 ):
+        #         x = [pos + i for i in range(-2,3) if modCalls.has_key( pos + i) and self.rawKinetics.has_key( pos + i) ]
+        #         y = [modCalls[j] for j in x]
+        #         if y.count('K') > 1:
+        #             tmp = [self.rawKinetics[j]['Ca5C'] for j in x if self.rawKinetics[j].has_key('Ca5C') ]
+        #             if len(tmp) > 0:
+        #                 lowest = min(tmp)
+        #                 for j in x:
+        #                     if self.rawKinetics[j].has_key('Ca5C'):
+        #                         if self.rawKinetics[j]['Ca5C'] > lowest:
+        #                             del modCalls[j]
+        #             
 
 
-            # if adjacent m5C calls are made by the LDA, select the one that has the lower LDA score (Ca5C)
-            # for pos in range(start, end):
-            #     if modCalls.has_key(pos) and self.rawKinetics.has_key(pos) and modCalls.has_key(pos+1) and self.rawKinetics.has_key(pos+1):
-            #         if self.rawKinetics[pos].has_key('Ca5C') and self.rawKinetics[pos+1].has_key('Ca5C'):
-            #             if self.rawKinetics[pos]['Ca5C'] < self.rawKinetics[pos+1]['Ca5C']:
-            #                 del modCalls[pos+1] 
-            #             else:
-            #                 del modCalls[pos] 
+        #     # if adjacent m5C calls are made by the LDA, select the one that has the lower LDA score (Ca5C)
+        #     # for pos in range(start, end):
+        #     #     if modCalls.has_key(pos) and self.rawKinetics.has_key(pos) and modCalls.has_key(pos+1) and self.rawKinetics.has_key(pos+1):
+        #     #         if self.rawKinetics[pos].has_key('Ca5C') and self.rawKinetics[pos+1].has_key('Ca5C'):
+        #     #             if self.rawKinetics[pos]['Ca5C'] < self.rawKinetics[pos+1]['Ca5C']:
+        #     #                 del modCalls[pos+1] 
+        #     #             else:
+        #     #                 del modCalls[pos] 
 
             
         return modCalls
@@ -460,18 +322,24 @@ class ModificationDecode(object):
 
 
             # FIXME:  Without this, currently, the identificationQv score is too low for many Ca5C sites
-            if self.useLDA:
-                if self.rawKinetics.has_key(pos):
-                    if self.rawKinetics[pos].has_key('Ca5C'):
-                        llr = -self.rawKinetics[pos]['Ca5C']
-                        qModScore = 100 * llr * log10e + 100*log1p(exp(-llr))*log10e
+            # if self.useLDA:
+            #     if self.rawKinetics.has_key(pos):
+            #         if self.rawKinetics[pos].has_key('Ca5C'):
+            #             llr = -self.rawKinetics[pos]['Ca5C']
+            #             qModScore = 100 * llr * log10e + 100*log1p(exp(-llr))*log10e
 
 
 
             if self.methylFractionFlag and self.rawKinetics.has_key(pos):
+
                 if self.rawKinetics[pos]["coverage"] > self.methylMinCov:
+
+                    # Instantiate mixture estimation methods:
+                    mixture = MixtureEstimationMethods(self.gbmModel.post, self.gbmModel.pre, self.rawKinetics, self.methylMinCov)
+
                     # Use modifiedMeanVectors and unmodifiedMeanVectors to calculate mixing proportion, and 95% CI limits.
-                    methylFracEst,methylFracLow,methylFracUpp = self.estimateMethylatedFractions(pos, unModifiedMeanVectors, modifiedMeanVectors, ModificationPeakMask[modNames[call]] )
+                    methylFracEst,methylFracLow,methylFracUpp = mixture.estimateMethylatedFractions(pos, unModifiedMeanVectors, modifiedMeanVectors, ModificationPeakMask[modNames[call]] )
+
                     qvModCalls[pos] = { 'modification' : modNames[call], 'QMod' : qModScore, 'LLR' : llr, 'Mask': maskPos, \
                                     FRAC: methylFracEst, FRAClow: methylFracLow, FRACup: methylFracUpp }
 
@@ -485,152 +353,6 @@ class ModificationDecode(object):
         return qvModCalls
 
 
-    # Return expected IPDs for a portion [start, end] of the sequence.
-
-    def getContextMeans(self, start, end, sequence):
-        meanVector = []
-        for pos in xrange(start, end+1):
-            ctx = sequence[(pos-self.pre):(pos + self.post + 1)].tostring()
-            if self.contextMeanTable.has_key(ctx):
-                meanVector.append(self.contextMeanTable[ctx])
-            else:
-                meanVector.append(self.gbmModel.getPredictions([ctx]))
-        return meanVector
-
-
-    # Return value of mixture model log likelihood function
-
-    def mixModelFn(self, p, a0, a1):
-        tmp = (1-p)*a0 + p*a1
-        return -np.log( tmp[ np.nonzero(tmp) ] ).sum()
-        # return -np.ma.log( tmp ).sum()
-
-
-    # Try to speed up calculation by avoiding a call to scipy.stats.norm.pdf() 
-
-    def replaceScipyNormPdf( self, data, mu ):
-        # return np.exp( -np.divide( data, mu) ) / mu
-        tmp = np.divide( data, mu )
-        return np.exp( np.subtract( tmp, np.power(tmp, 2) / 2.0 ) ) / mu 
-        # pdf for normal distribution: res = res / sqrt( 2 * pi ) (can factor out sqrt(2 * pi))
-
-
-    # Return optimum argument (mixing proportion) of mixture model log likelihood function.
-
-    def estimateSingleFraction(self, mu1, data, mu0, L ):
-        a0 = self.replaceScipyNormPdf( data, mu0 )
-        a1 = self.replaceScipyNormPdf( data, mu1 )
-        # if f'(0) < 0 (equ. a1/a0 < L), then f'(1) < 0 as well and solution p-hat <= 0
-        if np.divide(a1, a0).sum() <= L:
-            return 0.0
-        # if f'(1) > 0 (equ. a0/a1 < L), then f'(0) > 0 as well and solution p-hat >= 1
-        if np.divide(a0, a1).sum() <= L:
-            return 1.0
-        # unconstrained minimization of convex, single-variable function
-        res = fminbound(self.mixModelFn, 0.01, 0.99, args=(a0, a1), xtol=1e-02)
-        return res
-
-
-
-    # Try bias-corrected, accelerated quantiles for bootstrap confidence intervals
-
-    def bcaQuantile( self, estimate, bootDist, data, mu0, mu1, nSamples, n ):
-
-        tmp = sum( y <= estimate for y in bootDist ) / float(nSamples + 1)
-        if tmp > 0 and tmp < 1:
-
-            # bias correction
-            z0 = s.norm.ppf( tmp )
-
-            # acceleration
-            x = np.zeros(n)
-            for i in range(n):
-                x[i] = self.estimateSingleFraction(mu1, np.delete(data, i), mu0, n-1)
-            xbar = np.mean(x)
-            denom =  np.power( np.sum( np.power( x - xbar, 2) ), 1.5 )
-            if abs(denom) < 1e-4:
-                q1 = 2.5
-                q2 = 97.5
-            else:
-                a = np.divide( np.sum( np.power( x - xbar, 3) ),  denom ) / 6.0
-
-                # quantiles: (k1 and k2 are defined globally)
-                q1 = 100*s.norm.cdf( z0 + (z0 + k1)/(1 - a*(z0 + k1)) )
-                q2 = 100*s.norm.cdf( z0 + (z0 + k2)/(1 - a*(z0 + k2)) )
-
-        elif tmp == 0.0:
-            q1 = 0
-            q2 = 0
-
-        elif tmp == 1.0:
-            q1 = 100
-            q2 = 100
-
-        return (q1, q2)
-
-
-    # Bootstraps mix prop estimates to return estimate and simple bounds for 95% confidence interval
-
-    def bootstrap(self, pos, mu0, mu1, nSamples = 500):
-
-        if not self.rawKinetics.has_key( pos ):
-            return np.array( [ float('nan'), float('nan'), float('nan') ] )
-
-        res = np.zeros(3)
-        sample = self.rawKinetics[pos]["rawData"]
-        L = len(sample)
-        X = np.zeros(nSamples+1)
-        res[0] = self.estimateSingleFraction(mu1, sample, mu0, L)
-        X[nSamples] = res[0]
-
-        for i in range(nSamples):
-            bootstrappedSamples = sample[s.randint.rvs(0, L-1, size=L)]
-            X[i] = self.estimateSingleFraction(mu1, bootstrappedSamples, mu0, L)
-
-        q1,q2 = self.bcaQuantile( res[0], X, sample, mu0, mu1, (nSamples+1), L )
-        res[1] = np.percentile(X, q1)
-        res[2] = np.percentile(X, q2)
-        return res
-
-
-    # Returns [estimate, 95% CI lower bnd, 95% CI upper bound] using a weighted sum
-    # The hope is that this would work better for a multi-site signature, such as m5C_TET
-
-    def estimateMethylatedFractions(self, pos, meanVector, modMeanVector, maskPos ):
-
-        maskPos = np.array(maskPos)
-        L = len(maskPos)
-        if L == 0:
-            res = self.bootstrap(pos, meanVector[self.post], modMeanVector[self.post] )
-        else:
-            est = np.zeros(L)
-            low = np.zeros(L)
-            upp = np.zeros(L)
-            res = np.zeros(3)
-            wts = np.zeros(L)
-
-            # for offset in maskPos:
-            for count in range(L):
-                offset = maskPos[count]
-                mu0 = meanVector[ self.post + offset ]
-                mu1 = modMeanVector[ self.post + offset ]
-                if mu1 > mu0:
-                    k = self.bootstrap( (pos + offset), mu0, mu1 )
-                    wts[count] = k[0] * (mu1 - mu0)
-                    est[count] = k[0]
-                    low[count] = k[1]
-                    upp[count] = k[2]
-
-            if sum(wts) > 1e-3:
-                wts = wts/sum(wts)
-                res[0] = np.multiply(est, wts).sum()
-                res[1] = np.multiply(low, wts).sum()
-                res[2] = np.multiply(upp, wts).sum()
-
-        # print str(res)
-        return res
-
-
 
     def scoreRegion(self, start, end, sequence):
 
@@ -641,6 +363,7 @@ class ModificationDecode(object):
                 sc += self.scores[pos][ctx]
 
         return sc
+
 
 
     def getRegionScores(self, start, end, sequence):
@@ -672,57 +395,5 @@ class ModificationDecode(object):
 
     def compareStates(self, current, prev):
         return current[0:-1] == prev[1:]
-
-
-
-    # Everything below here is unused for now:
-
-
-    # Return second derivative of mixture model log likelihood function - unused for now
-    def mixModelFnPrime2(self, p, a0, a1):
-        tmp = np.square( (1-p)*a0 + p*a1 )
-        nonzero_indices = np.nonzero(tmp)
-        return np.divide( np.square(a1 - a0)[nonzero_indices], tmp[nonzero_indices] ).sum()
-
-
-    # Return third derivative of mixture model log likelihood function - unused for now
-    def mixModelFnPrime3(self, p, a0, a1):
-        tmp = np.power( (1-p)*a0 + p*a1, 3 )
-        nonzero_indices = np.nonzero(tmp)
-        return -np.divide( np.power(a1 - a0, 3)[nonzero_indices], tmp[nonzero_indices] ).sum()
-
-
-    # Try removing very large values before case resampling for bootstrap estimation - unused for now
-    def processSample( self, sample ):
-        q1 = np.percentile(sample, 25)
-        q2 = np.percentile(sample, 75)
-        iqr = 1.5*(q2 - q1)
-        uif = q2 + iqr
-        lif = q1 - iqr
-        def removeBoxplotOutliers(x):
-            if (x > lif) and (x < uif):
-                return x
-        return filter(removeBoxplotOutliers, sample)
-
-
-    # Return derivative of mixture model log likelihood function -- unused for now
-    def mixModelFnPrime(self, p, a0, a1):
-        tmp = (1-p)*a0 + p*a1
-        nonzero_indices = np.nonzero(tmp)
-        return -np.divide( (a1 - a0)[nonzero_indices], tmp[nonzero_indices] ).sum()
-
-
-    # unconstrained minimization of convex, single-variable function - unused for now
-    # much slower than fminbound
-    def homeMadeMinimization( self, a0, a1, low, up, xtol = 1e-02, maxIters = 500 ):
-        nIters = 0
-        while (up - low) > xtol and nIters < maxIters:
-            p0 = (up - low)/2.0
-            if self.mixModelFnPrime(p0, a0, a1) <= 0:
-                low = p0
-            else:
-                up = p0
-            nIters += 1
-        return p0
